Adding services using "criteria query"
