+++ pom.xml
@@ -36,7 +36,17 @@
 			<artifactId>spring-data-jpa</artifactId>
 			<version>1.9.0.RELEASE</version>
 		</dependency> -->
-
+	    <dependency>
+	        <groupId>log4j</groupId>
+	        <artifactId>log4j</artifactId>
+	        <version>1.2.16</version>
+	    </dependency>
+	
+	    <dependency>
+	        <groupId>org.slf4j</groupId>
+	        <artifactId>slf4j-log4j12</artifactId>
+	        <version>1.6.4</version>
+	    </dependency>
 	</dependencies>
 	<scm>
 		<developerConnection>Christian Monzón</developerConnection>
